daemon/gpiod.c: daemon core
This initial commit marks the first step towards building out a
simple daemon for libgpiod.

There is a truck ton to add to this project, but as it currently
stands the project provides a skeleton daemon that can be easily
extended upon in order to introduce additional functionality. The core
idea behind this daemon is to have as little dependencies on other
system components as possible and hence why it takes the SysV daemon
bootstrapping approach.

Signed-off-by: Matthew Bobrowski <matthew@bobrowski.net>